StringProcessingTest: Move tests
Moves the tests to the same scope like in coalib.
