+++ coalib/tests/parsing/StringProcessing/SearchInBetweenTest.py
@@ -2,24 +2,25 @@ import sys
 import unittest
 
 sys.path.insert(0, ".")
-from coalib.tests.parsing.StringProcessingTest import StringProcessingTest
-from coalib.parsing.StringProcessing import unescaped_search_in_between
+from coalib.tests.parsing.StringProcessing.StringProcessingTest import (
+    StringProcessingTest)
+from coalib.parsing.StringProcessing import search_in_between
 
 
-class UnescapedSearchInBetweenTest(StringProcessingTest):
+class SearchInBetweenTest(StringProcessingTest):
     bs = StringProcessingTest.bs
 
     test_basic_pattern = "'"
     test_basic_expected_results = [
         [r"escaped-escape:        \\ "],
-        [r"escaped-quote:         \' "],
+        [r"escaped-quote:         " + bs],
         [r"escaped-anything:      \X "],
         [r"two escaped escapes: \\\\ "],
-        [r"escaped-quote at end:   \'"],
+        [r"escaped-quote at end:   " + bs],
         [r"escaped-escape at end:  " + 2 * bs],
         [r"str1", r"str2"],
-        [r"str1", r"str2"],
-        [r"str1", r"str2"],
+        [r"        ", r" out2 "],
+        [r"      ", r" out2 "],
         [r"str1", r"str2"],
         [r"str1", r"str2"],
         [r"str1", r"str2"],
@@ -30,12 +31,12 @@ class UnescapedSearchInBetweenTest(StringProcessingTest):
         [],
         []]
 
-    # Test the basic unescaped_search_in_between() functionality.
+    # Test the basic search_in_between() functionality.
     def test_basic(self):
         expected_results = self.test_basic_expected_results
 
         self.assertResultsEqual(
-            unescaped_search_in_between,
+            search_in_between,
             {(self.test_basic_pattern,
               self.test_basic_pattern,
               test_string,
@@ -47,14 +48,14 @@ class UnescapedSearchInBetweenTest(StringProcessingTest):
              for use_regex in [True, False]},
             list)
 
-    # Test the unescaped_search_in_between() while varying the max_match
+    # Test the search_in_between() while varying the max_match
     # parameter.
     def test_max_match(self):
         search_pattern = self.test_basic_pattern
         expected_master_results = self.test_basic_expected_results
 
         self.assertResultsEqual(
-            unescaped_search_in_between,
+            search_in_between,
             {(search_pattern,
               search_pattern,
               test_string,
@@ -68,21 +69,21 @@ class UnescapedSearchInBetweenTest(StringProcessingTest):
              for use_regex in [True, False]},
             list)
 
-    # Test the unescaped_search_in_between() function with different regex
+    # Test the search_in_between() function with different regex
     # patterns.
     def test_regex_pattern(self):
         expected_results = [
             [r""],
             [r"c"],
             [r"c", r"bc\+'**'"],
-            [r"\13q4ujsabbc"],
+            [r""],
             [r"\\13q4ujsabbc\+'**'ac", r"."],
             [r"", r"", r"", r"", r"", r"c\+'**'", r"", r"", r"-"],
             [r"cba###\\13q4ujs"],
-            []]
+            [r"3q4ujsabbc" + self.bs]]
 
         self.assertResultsEqual(
-            unescaped_search_in_between,
+            search_in_between,
             {(pattern,
               pattern,
               self.multi_pattern_test_string,
@@ -93,12 +94,12 @@ class UnescapedSearchInBetweenTest(StringProcessingTest):
                                         expected_results)},
             list)
 
-    # Test the unescaped_search_in_between() function for its
+    # Test the search_in_between() function for its
     # remove_empty_matches feature.
     def test_auto_trim(self):
         expected_results = [
             [],
-            [r"\\\\\;\\#", r"+ios"],
+            [5 * self.bs, r"\\\'", self.bs, r"+ios"],
             [r"2", r"4", r"6"],
             [r"2", r"4", r"6"],
             [],
@@ -108,7 +109,7 @@ class UnescapedSearchInBetweenTest(StringProcessingTest):
             [r"a"]]
 
         self.assertResultsEqual(
-            unescaped_search_in_between,
+            search_in_between,
             {(self.auto_trim_test_pattern,
               self.auto_trim_test_pattern,
               test_string,
@@ -120,14 +121,13 @@ class UnescapedSearchInBetweenTest(StringProcessingTest):
              for use_regex in [True, False]},
             list)
 
-    # Test the unescaped_search_in_between() function for its use_regex
-    # parameter.
+    # Test the search_in_between() function for its use_regex parameter.
     def test_disabled_regex(self):
-        search_pattern = r"'()?"
+        search_pattern = r"\'"
         expected_results = [[] for x in range(len(self.test_strings))]
 
         self.assertResultsEqual(
-            unescaped_search_in_between,
+            search_in_between,
             {(search_pattern,
               search_pattern,
               test_string,
@@ -139,19 +139,20 @@ class UnescapedSearchInBetweenTest(StringProcessingTest):
              for auto_trim in [True, False]},
             list)
 
-    # Test the unescaped_search_in_between() function using the test-strings
-    # specific for search-in-between functions.
+    # Test the search_in_between() function using the test-strings specific
+    # for search-in-between functions.
     def test_extended(self):
         expected_results = [
             [r"", r"This is a word", r"(in a word"],
             [r"((((((((((((((((((1"],
             [r"do (it ", r"", r"hello."],
-            [r"", r"This\ is a word\)and((in a\\\ word\\\\\) another \)"],
-            [r"((((\\\(((((((((((1"],
-            [r"it ", r"", r"hello."]]
+            [r"", r"This\ is a word" + self.bs,
+                r"(in a\\\ word" + 5 * self.bs],
+            [r"\(\((((((\\\(((((((((((1"],
+            [r"do (it ", r"", r"hello."]]
 
         self.assertResultsEqual(
-            unescaped_search_in_between,
+            search_in_between,
             {(begin_pattern,
               end_pattern,
               test_string,
