+++ coalib/tests/parsing/StringProcessing/StringProcessingTest.py
