+++ coalib/tests/parsing/StringProcessing/TrimEmptyMatchesTest.py
